# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Lottery Backend
  description: "Backend учебной лотереи: регистрация пользователей, создание тиражей, продажа билетов и определение победителя."
  version: 1.0.0
  contact:
    name: Lottery Team
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: "Создаёт нового пользователя с ролью USER. Повторяющийся логин возвращает 409."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              login: alice
              password: supersecret123
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                id: 1
                login: alice
                message: "User registered successfully"
        "400":
          description: Невалидный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseBody"
              example:
                code: VALIDATION_FAILED
                message: "login must not be blank"
        "409":
          description: Логин уже занят
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseBody"
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Авторизация пользователя
      description: "Аутентифицирует пользователя и возвращает Bearer-токен (UUID). Если уже авторизован — возвращает существующий токен."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              login: alice
              password: supersecret123
      responses:
        "200":
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                message: "Login successful"
                token: "550e8400-e29b-41d4-a716-446655440000"
                id: 1
                login: alice
                role: USER
        "401":
          description: Неверный логин/пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseBody"
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Выход из системы
      description: "Удаляет Bearer-токен текущего пользователя из сессии."
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Успешный выход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
              example:
                message: "Logout successful"
        "401":
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseBody"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Bearer-токен в формате UUID, выдаётся `POST /auth/login`."
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 50
          description: "Логин пользователя. Должен быть уникальным."
          example: alice
        password:
          type: string
          minLength: 8
          description: "Пароль пользователя. Не возвращается в ответах."
          example: supersecret123
      required:
        - login
        - password
    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required:
        - login
        - password
    RegisterResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
        message:
          type: string
      required:
        - id
        - login
        - message
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
          description: "Bearer-токен (UUID). Передавать в заголовке `Authorization: Bearer <token>`."
        id:
          type: integer
          format: int64
        login:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
      required:
        - message
        - token
        - id
        - login
        - role
    LogoutResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ErrorResponseBody:
      type: object
      properties:
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - CONFLICT
            - NOT_FOUND
            - UNAUTHORIZED
            - FORBIDDEN
        message:
          type: string
      required:
        - code
        - message